add concole command and refactoring
