+++ src/AppBundle/Entity/Coach.php
@@ -7,9 +7,39 @@ use Doctrine\ORM\Mapping as ORM;
 /**
  * User
  *
- * @ORM\Entity(repositoryClass="AppBundle\Entity\CoachRepository")
+ * @ORM\Entity(repositoryClass="AppBundle\Repository\CoachRepository")
  */
 class Coach extends User
 {
     const ROLENAME = 'coach';
+
+    /**
+     *@ORM\ManyToOne(targetEntity="Team", inversedBy="coaches")
+     */
+    protected $team;
+
+    /**
+     * Set product
+     *
+     * @param \AppBundle\Entity\Team $team
+     *
+     * @return User
+     */
+    public function setTeam(Team $team = null)
+    {
+        $this->team = $team;
+
+        return $this;
+    }
+
+    /**
+     * Get product
+     *
+     * @return \AppBundle\Entity\Team
+     */
+    public function getTeam()
+    {
+        return $this->team;
+    }
+
 }
