--- src/AppBundle/Model/Country.php
@@ -1,12 +1,0 @@
-<?php
-
-namespace AppBundle\Model;
-
-
-class Country
-{
-    public $name;
-    public $slug;
-    public $image;
-    public $info;
-}
