--- src/AppBundle/Model/Game.php
@@ -1,17 +1,0 @@
-<?php
-
-namespace AppBundle\Model;
-
-
-class Game
-{
-    public $id;
-    public $team1Name;
-    public $team1Slug;
-    public $team1Score;
-    public $team1Image;
-    public $team2Name;
-    public $team2Slug;
-    public $team2Score;
-    public $team2Image;
-}
