--- src/AppBundle/Model/Team.php
@@ -1,14 +1,0 @@
-<?php
-
-namespace AppBundle\Model;
-
-
-class Team
-{
-    public $name;
-    public $slug;
-    public $country_slug;
-    public $country_image;
-    public $info;
-    public $image;
-}
