--- src/AppBundle/Model/User.php
@@ -1,12 +1,0 @@
-<?php
-
-namespace AppBundle\Model;
-
-
-class User
-{
-    public $name;
-    public $lastName;
-    public $info;
-    public $id;
-}
