+++ src/AppBundle/Repository/BracketRepository.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace AppBundle\Entity;
+namespace AppBundle\Repository;
 
 /**
  * BracketRepository
