+++ src/AppBundle/Repository/CoachRepository.php
@@ -1,13 +1,13 @@
 <?php
 
-namespace AppBundle\Entity;
+namespace AppBundle\Repository;
 
 /**
- * CountryRepository
+ * CoachRepository
  *
  * This class was generated by the Doctrine ORM. Add your own custom
  * repository methods below.
  */
-class CountryRepository extends \Doctrine\ORM\EntityRepository
+class CoachRepository extends \Doctrine\ORM\EntityRepository
 {
 }
