+++ src/AppBundle/Repository/CountryRepository.php
@@ -1,13 +1,13 @@
 <?php
 
-namespace AppBundle\Entity;
+namespace AppBundle\Repository;
 
 /**
- * RoleRepository
+ * CountryRepository
  *
  * This class was generated by the Doctrine ORM. Add your own custom
  * repository methods below.
  */
-class RoleRepository extends \Doctrine\ORM\EntityRepository
+class CountryRepository extends \Doctrine\ORM\EntityRepository
 {
 }
