+++ src/AppBundle/Repository/PlayerRepository.php
@@ -1,13 +1,13 @@
 <?php
 
-namespace AppBundle\Entity;
+namespace AppBundle\Repository;
 
 /**
- * CoachRepository
+ * PlayerRepository
  *
  * This class was generated by the Doctrine ORM. Add your own custom
  * repository methods below.
  */
-class CoachRepository extends \Doctrine\ORM\EntityRepository
+class PlayerRepository extends \Doctrine\ORM\EntityRepository
 {
 }
