+++ src/AppBundle/Repository/RoleRepository.php
@@ -1,13 +1,13 @@
 <?php
 
-namespace AppBundle\Entity;
+namespace AppBundle\Repository;
 
 /**
- * PlayerRepository
+ * RoleRepository
  *
  * This class was generated by the Doctrine ORM. Add your own custom
  * repository methods below.
  */
-class PlayerRepository extends \Doctrine\ORM\EntityRepository
+class RoleRepository extends \Doctrine\ORM\EntityRepository
 {
 }
