+++ src/AppBundle/Repository/TeamRepository.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace AppBundle\Entity;
+namespace AppBundle\Repository;
 
 /**
  * TeamRepository
